requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.2.1
django-crispy-forms==2.0
django-isbn-field==0.5.3
gunicorn==20.1.0
Pillow==9.5.0
python-stdnum==1.18
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
